flag: Move deprecated skip verify to flag
This is to keep related insecure flags close to each other.

Also, before releasing 1.0 we have to remove just one file and one
function call in rkt.go.
